Use env vars for server URLs
At this commit it seems that Cypress tests only work when BASE_URL is set to
the "Network" IP, not localhost:

  App running at:
  - Local:   http://localhost:8080/
  - Network: http://10.1.2.2:8080/

I don't know why that is yet. It also should be possible to set the
CORS response headers in cors.rs to "$UI_SERVER_URL", rather than "*".

Notes:

Vue
===

The Vue build process reads the .env file and arranges that any
variables in there named VUE_APP_* will be available in process.env at
run time in the client-side javascript. The prefix is not stripped. It
does not use standard OS environment variables; it gets variables by
reading the .env file.

The .env file is not a shell script. For example, you can't reuse
variables using $MY_VARIABLE. It's just a sequence of key-value pairs,
one per line.

We need the Vue code at run time to know the URL of the backend JSON
API server. So we make an entry for it in .env with the name
VUE_APP_SERVER_URL.

Cypress
=======

Cypress is client-side javascript executing _in the page_ that you
tell it to visit. When Cypress is executing, we require two things to
be true:

1. Cypress knows the URL of the UI server so that it can visit the
   page it is supposed to be testing. For example, this might be a dev
   server with live-reloading used during javascript development, or
   it might be the backend server, which is serving the frontend from
   static files, which were created by `npm build`.

2. process.env.VUE_APP_SERVER_URL exists in the client-side javascript
   run time environment, and holds the URL of the backend JSON API
   server. This is necessary because Cypress is testing the Vue code and the
   Vue code makes requests to process.env.VUE_APP_SERVER_URL as described
   above; the fact that Cypress code is also executing client-side
   doesn't change this.

Now, Cypress does not read the .env file; it uses traditional OS
environment variables. So the Makefile arranges that the cypress
process has environment variables injected into it from the .env file.
Furthermore, Cypress only makes environment variables named CYPRESS_*
available in the client-side run time environment. However, unlike
Vue, Cypress _does_ strip the prefix.

So, to satisfy requirement (1), we create an entry in .env named
CYPRESS_BASE_URL. Cypress automatically prefixes any routes with this
URL. The result is that Cypress test code looks like cy.visit("/"),
which is really cy.visit(`${CYPRESS_BASE_URL}/`).

Requirement (2) is satisfied in the same way as it is when we are not
using Cypress by the VUE_APP_SERVER_URL entry in .env.

Backend
=======

Finally, the backend need to know the address of the server that has
served javascript making ajax requests, in order that the backend can
send CORS response headers allowing such requests to proceed. For
that, we have an entry in .env named UI_SERVER_URL. The backend code
doesn't read the .env file; the Makefile creates standard OS
environment variables from it when starting the backend server process.