Switch back to level (with modal option) for game form fields
The table (if it has multiple tds) cannot straightforwardly be used
with the modal option due to the requirement for a single top-level
element in GameFormField.vue. The reason for using a table was to
handle vertical alignment of multiple tds. But now that the field
labels are in the input placeholder, we do not have multiple tds, and
we could use either level or table.