Hack / bug-fix: style input on return from modal
GameFormFieldContainer.vue uses v-if...else to render either the modal
or non-modal version of the autocomplete. I think that, since it is
v-if, and not v-show, when the modal closes, the non-modal version is
created and mounted from scratch. The answer (from the modal input) is
already there and it seems that calling styleInputAccordingToAnswer in
the mounted hook works, as this commit does.

My first thought was to use the modal `close` or `after-leave`
events (https://buefy.org/documentation/modal/#api-view) but (a) I'm
not clear how to set event handlers on a buefy modal created via a
template (see https://github.com/buefy/buefy/issues/1947), and (b)
while I can create a modal programmatically in
GameFormFieldContainer.vue (something like below), I don't know how to
make it show the game form field component (since this is available as
a slot in GameFormFieldContainer.vue).

  watch: {
    isModal() {
      if (this.isModal) {
        let modal = this.$buefy.modal.open({
          parent: this,
          fullScreen: true,
          content: "hello",
          // component: this.$children[0], this.$slots ?
        });
        modal.$on("close", () => console.log("modal close!!"));
      }
    },
  },