DNW: Work around browser blocking of audio.play()
This doesn't actually work: it seems that Chrome refuses to play the
audio if it originates from an HTML tag subject to Vue's v-if or
v-show directives.

The previous commit seemed to autoplay on everything except Chrome
desktop (it autoplayed on Safari & Firefox desktop, and Chrome &
Safari iPhone). However, Chrome desktop was being very stubborn: I was
unable to ever make the audio.play() call start playback, whether
invoked as in the code, or from dev tools after having already started
and paused the audio with genuine user interaction.

With the use of an explicit <audio> tag in this commit, I think what
happens is that the Audio object that is created programmatically
triggers download of the audio data in advance, as we want. Then, we
don't actually mount that Audio object in the DOM, but rather use the
one from the <audio> tag, with the same URL. I think that the
browser (Chrome desktop, at least) re-uses the already-downloaded data
for that URL and thus we do get a fast transition (which was the sole
motivation for the changes in the previous commit and this one).
Unfortunately, the transition is no longer instantaneous as it was on
the previous commit.